AuditMind/requirements.txt
numpy==1.24.4
opencv-python==4.10.0.84
paddlepaddle==3.0.0
paddleocr==3.0.1
pdf2image==1.16.3
pillow==10.0.0
python-magic-bin==0.4.14
pdfplumber==0.11.0
pandas==2.1.0
openpyxl==3.1.2
requests==2.31.0